SAS-343 Movidos tests de play() para que se encuentren juntos los correspondientes a game_effect (play + get_defense_card)
